REVIEW: comments addressed
