+++ src/components/cl/basic/cl_basic.c
@@ -42,6 +42,6 @@ UCC_CLASS_DEFINE_DELETE_FUNC(ucc_cl_basic_context_t, ucc_base_context_t);
 UCC_CLASS_DEFINE_NEW_FUNC(ucc_cl_basic_team_t, ucc_base_team_t,
                           ucc_base_context_t *, const ucc_base_team_params_t *);
 
-extern ucc_status_t ucc_cl_basic_team_create_test(ucc_base_team_t *cl_team);
-extern ucc_status_t ucc_cl_basic_team_destroy(ucc_base_team_t *cl_team);
+ucc_status_t ucc_cl_basic_team_create_test(ucc_base_team_t *cl_team);
+ucc_status_t ucc_cl_basic_team_destroy(ucc_base_team_t *cl_team);
 UCC_CL_IFACE_DECLARE(basic, BASIC);
