+++ src/components/cl/basic/cl_basic_team.c
@@ -1,5 +1,5 @@
 /**
- * Copyright (C) Mellanox Technologies Ltd. 2020.  ALL RIGHTS RESERVED.
+ * Copyright (C) Mellanox Technologies Ltd. 2020-2021.  ALL RIGHTS RESERVED.
  *
  * See file LICENSE for terms.
  */
@@ -17,14 +17,14 @@ UCC_CLASS_INIT_FUNC(ucc_cl_basic_team_t, ucc_base_context_t *cl_context,
     UCC_CLASS_CALL_SUPER_INIT(ucc_cl_team_t, &ctx->super);
     status = UCC_TL_CTX_IFACE(ctx->tl_ucp_ctx)
                  ->team.create_post(&ctx->tl_ucp_ctx->super, params, &b_team);
-    if (UCC_OK == status) {
-        self->tl_ucp_team = ucc_derived_of(b_team, ucc_tl_team_t);
-        cl_info(cl_context->lib, "posted cl team: %p", self);
-    } else {
+    if (UCC_OK != status) {
         self->tl_ucp_team = NULL;
         cl_error(cl_context->lib, "tl ucp team create post failed");
+        return status;
     }
-    return status;
+    self->tl_ucp_team = ucc_derived_of(b_team, ucc_tl_team_t);
+    cl_info(cl_context->lib, "posted cl team: %p", self);
+    return UCC_OK;
 }
 
 UCC_CLASS_CLEANUP_FUNC(ucc_cl_basic_team_t)
