+++ src/components/tl/ucp/Makefile.am
@@ -9,7 +9,9 @@ sources =            \
 	tl_ucp_context.c \
 	tl_ucp_team.c    \
 	tl_ucp_ep.h      \
-    tl_ucp_ep.c
+	tl_ucp_ep.c      \
+	tl_ucp_addr.h    \
+	tl_ucp_addr.c
 
 module_LTLIBRARIES = libucc_tl_ucp.la
 libucc_tl_ucp_la_SOURCES  = $(sources)
