+++ src/components/tl/ucp/tl_ucp.h
@@ -1,5 +1,5 @@
 /**
- * Copyright (C) Mellanox Technologies Ltd. 2020.  ALL RIGHTS RESERVED.
+ * Copyright (C) Mellanox Technologies Ltd. 2020-2021.  ALL RIGHTS RESERVED.
  *
  * See file LICENSE for terms.
  */
@@ -24,7 +24,7 @@ typedef struct ucc_tl_ucp_lib_config {
 
 typedef struct ucc_tl_ucp_context_config {
     ucc_tl_context_config_t super;
-    int                     preconnect;
+    uint32_t                preconnect;
 } ucc_tl_ucp_context_config_t;
 
 typedef struct ucc_tl_ucp_lib {
@@ -40,21 +40,20 @@ typedef struct ucc_tl_ucp_context {
     ucp_worker_h     ucp_worker;
     size_t           ucp_addrlen;
     ucp_address_t   *worker_address;
-    int              preconnect;
+    uint32_t         preconnect;
     ucc_tl_ucp_ep_close_state_t ep_close_state;
 } ucc_tl_ucp_context_t;
 UCC_CLASS_DECLARE(ucc_tl_ucp_context_t, const ucc_base_context_params_t *,
                   const ucc_base_config_t *);
 
 typedef struct ucc_tl_ucp_team {
-    ucc_tl_team_t super;
-    ucc_status_t  status;
-    int           context_ep_storage; /*< The flag indicates whether ucp endpoints
-                                          are stored on the ucc_tl_ucp_context or
-                                          are they created per-team.
-                                          This optimization is only possible when
-                                          user provides the necessary rank mappings
-                                          team_rank->context_rank. */
+    ucc_tl_team_t              super;
+    ucc_status_t               status;
+    int                        context_ep_storage; /*< The flag
+              indicates whether ucp endpoints are stored on the
+              ucc_tl_ucp_context or are they created per-team.
+              This optimization is only possible when user provides
+              the necessary rank mappings team_rank->context_rank. */
     ucp_ep_h                  *eps;
     int                        size;
     int                        rank;
@@ -66,17 +65,6 @@ UCC_CLASS_DECLARE(ucc_tl_ucp_team_t, ucc_base_context_t *,
 typedef struct ucc_tl_ucp_req {
     ucc_status_t status;
 } ucc_tl_ucp_req_t;
-
-typedef struct ucc_tl_ucp_addr_storage {
-    void                 *oob_req;
-    size_t                max_addrlen;
-    int                   state;
-    int                   is_ctx;
-    size_t               *addrlens;
-    void                 *addresses;
-    ucc_team_oob_coll_t   oob;
-    ucc_tl_ucp_context_t *ctx;
-} ucc_tl_ucp_addr_storage_t;
 
 #define UCC_TL_UCP_TEAM_CTX(_team)                                             \
     (ucc_derived_of((_team)->super.super.context, ucc_tl_ucp_context_t))
